refactor: 抽象应用管理功能为独立的 app 模块
主要变更：
- 创建独立的 app 功能模块 (src/features/app/)
  - tools.ts: 5个应用管理工具
  - handlers.ts: 应用选择和列表处理器
  - api.ts: 应用相关的 API 调用
  - index.ts: 模块定义和注册

- 从 leaderboard 模块分离应用操作
  - 移除 leaderboard/api.ts 中的应用相关函数和接口 (-200行)
  - 移除 leaderboard/tools.ts 中的应用管理工具 (-79行)
  - 更新 leaderboard/handlers.ts 从 app 模块导入

- 简化 core/handlers/appHandlers.ts
  - 改为兼容层，仅导出 app/handlers.ts 的函数
  - 添加 @deprecated 标记

- 更新服务器注册
  - 在 server.ts 中注册 app 模块
  - 启动日志显示: app (5 tools, 0 resources)

架构优势：
- 清晰的模块边界和职责分离
- 应用管理功能可被其他模块复用
- 符合单一职责和开闭原则
- 为未来功能扩展奠定基础

文档：
- 添加 REFACTOR_APP_MODULE.md 详细说明重构内容
